arreglamos parametro en la cabecera y hacemos que busque el mejor para hacer la salida para graficas, tambien eliminamos la BL de explotar, no conseguimos mucho con esto

+++ include/PAR.h
@@ -94,7 +94,7 @@ class PAR{
 
 		std::pair<std::vector<PAR::Cluster>, double> algoritmo_propio(const int MAX_EVAL, const int TAM_POB_INI,
 																						  const double PROB_CAMBIAR_GEN, const double PORCENTAJE_EXPLORAR,
-																					  	  const double PORCENTAJE_MUTAR, const int PORCENTAJE_INTERCAMBIAR,
+																					  	  const double PORCENTAJE_MUTAR, const double PORCENTAJE_INTERCAMBIAR,
 																						  const bool SALIDA = true);
 
 		int calcular_infactibilidad() const;
